Update status to rejected for Devcon 6 (#47)
Hi guys,

Many thanks for all of the detailed responses and explanations. We’ve been deliberating as a team on the various proposals we’ve received, and sadly we’ve decided not to move forward with GET Protocol and GUTS Ticketing for *this coming* Devcon. 

We do however really appreciate the proposal and the time you spent to walk us through your solution. It’s impressive what you guys have built and I think it suits some particular needs of the event ecosystem very well. 

For Devcon however a few of our values directed us to pursue other options:

— We try to promote FOSS software and mentality whenever we can, and some of our other options satisfy this goal better. We understand the GET Protocol is open source, but the GUTS ticketing app being closed source is a downside from Devcon’s values perspectives. 

— Requiring phone numbers is a no go for Devcon. We appreciate your explanations of how it could work without phone numbers but it seems like it would be a large deviation from what’s currently implemented. I understand it’s an integral part of the no-scalping features (as explained in your principles as well), and for certain more-mainstream use cases it probably makes more sense.

— Given the nature of our attendees and the fact that it’s a conference heavily attended by developers, we want to give the users greater functionality and power over their blockchain based asset and give other third party developers the ability to interact and extend, rather than hide that capability behind a smooth UX. In a mainstream event, most people wouldn’t care about interacting with their blockchain based asset, NFT, etc. (not yet at least 😉) but at Devcon, many people would. 

— We really appreciate your willingness to adapt, modify, and accommodate feature requests, but much of what we discussed as needs look like they’ll need to be custom built and that introduces more unknowns than we’re comfortable with. 

— Other proposals are able to integrate with Pretix, which we are already comfortable and satisfied with (and is open source). 

Those compiled with a few other reasons led us to pass on this proposal. I do believe though that what you are building is in general great, and for other use cases would make a lot of sense. I also want to communicate that this decision does not remove GET Protocol and GUTS from consideration for future Devcons -- it sounds like some of the features we discussed as needs are already in the works and on your roadmap.

All in all, many thanks for the time and the proposal. Our team would have not known this project existed without you reaching out. Best of luck building out the project, and hope to see you all at Devcon!

Best,
Skylar & Devcon team